requirements.txt
pandas
openpyxl
sqlalchemy
pymysql
python-dotenv
xlsxwriter
jupyter
cryptography